+++ recursion/snip/fibonacci_loop.rs
@@ -1,4 +1,4 @@
-fn fibonacci_loop(n: u32) -> u32 {
+fn fibonacci_loop(n: u64) -> u64 {
     let (mut a, mut b) = (0, 1);
     for _ in 0..n {
         (a, b) = (b, a + b);
