+++ recursion/snip/fibonacci_rec.rs
@@ -1,5 +1,5 @@
-fn fibonacci_rec(n: u32) -> u32 {
-    fn imp(a: u32, b: u32, m: u32) -> u32 {
+fn fibonacci_rec(n: u64) -> u64 {
+    fn imp(a: u64, b: u64, m: u64) -> u64 {
         match m {
             0 => a,
             _ => imp(b, a + b, m - 1),
